캐릭터 정보창 도구 슬롯 구현
캐릭터 정보창 도구 슬롯에 도구 아이템 장착/해제 구현

+++ Assets/Scripts/Gatherable.cs
@@ -90,7 +90,7 @@ public class Gatherable : MonoBehaviour
     {
         if (other.tag.Equals("Tool") && Player.instance.IsAttacking)
         {
-            if (gatherableType == other.GetComponent<Tool>().toolType - 1 || other.GetComponent<Tool>().toolType == (int)Tool.ToolType.Hand)
+            if (gatherableType == (int)other.GetComponent<Tool>().toolType - 1 || other.GetComponent<Tool>().toolType == (int)Tool.ToolType.Hand)
             {
                 if (Player.instance.Hitable)
                 {
